Refactor data-table to be more generic
