+++ services/frontend/src/routes/(app)/test/(components)/index.ts
@@ -1,10 +1,10 @@
-export { default as DataTableCheckbox } from './data-table-checkbox.svelte';
-export { default as DataTableTitleCell } from './data-table-title-cell.svelte';
-export { default as DataTableStatusCell } from './data-table-status-cell.svelte';
-export { default as DataTableRowActions } from './data-table-row-actions.svelte';
-export { default as DataTablePriorityCell } from './data-table-priority-cell.svelte';
-export { default as DataTableColumnHeader } from './data-table-column-header.svelte';
+export { default as DataTableCheckbox } from '$lib/components/data-table/data-table-checkbox.svelte';
+export { default as DataTableTitleCell } from './data-table-cell-title.svelte';
+export { default as DataTableStatusCell } from './data-table-cell-status.svelte';
+export { default as DataTableRowActions } from './data-table-cell-actions.svelte';
+export { default as DataTablePriorityCell } from './data-table-cell-priority.svelte';
+export { default as DataTableColumnHeader } from '$lib/components/data-table/data-table-column-header.svelte';
 export { default as DataTableToolbar } from './data-table-toolbar.svelte';
-export { default as DataTablePagination } from './data-table-pagination.svelte';
-export { default as DataTableViewOptions } from './data-table-view-options.svelte';
-export { default as DataTableFacetedFilter } from './data-table-faceted-filter.svelte';
+export { default as DataTablePagination } from '$lib/components/data-table/data-table-pagination.svelte';
+export { default as DataTableViewOptions } from '../../../../lib/components/data-table/data-table-view-options.svelte';
+export { default as DataTableFacetedFilter } from '$lib/components/data-table/data-table-faceted-filter.svelte';
